requirements.txt

numpy
scipy
pandas
scikit-learn
mne
pyyaml
tensorflow
